Fix compute_score
torchbenchmark isn't installed in a proper way, so it doesn't work
to use relative imports from a script with a main function that is
nested inside the torchbenchmark package. work around this for now.

make machine_config script agnostic to cpu core pinning when run
as a script, since pinning is  checked by conftest during benchmarking
and it's pointless to make the user pin the configuration script.