Move LTC enablement into each model
ModelTask runs test in subprocess instead of the current process, which run.py does.

In order to have an E2E test trail for all run.py/test.py/test_bench.py, this patch moves the enablement logic down to the model. And leaves a TODO to generalize the approach.