Fix the import issue of speech_transformer. (#542)
Summary:
When adding the speech_transformer model, we modified the syspath import path.
This is actually undesired because it will also affect the import path of other models. Although in CI environment we are running models in separate processes, this could still cause problem when locally running with `run.py`.

Therefore, here we are moving the model to a subdirectory `speech_transformer`, and change all the import styles to relative imports.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/542

Reviewed By: aaronenyeshi

Differential Revision: D32139001

Pulled By: xuzhao9

fbshipit-source-id: 5d94d327c3c897da36b1739d5854f479eaf8fe4d