Fix bug where unit tests are skipped when get_module raises NotImplemented
Since we added the check_device() to each of the unit tests: train, eval,
example, and check_device, some unit tests are skipped when get_module()
is NotImplemented. This is because we catch NotImplemented in test.py
instead of check_device. We should catch it here, and not skip the
rest of the test.