Only load the module specified in run.py argument (#577)
Summary:
When running `run.py`, it tries to read and load every model in the `models` directory even though we only need to run one specific model. In practice this often causes problems such as requiring to install `torchvision`/`torchtext` which are not required by the model we are attempting to run.

`run.py` should only load the specific model we are trying to run, instead of loading all the models.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/577

Reviewed By: aaronenyeshi

Differential Revision: D32653256

Pulled By: xuzhao9

fbshipit-source-id: 8b47dae2c45eebdf61752ad456c8bbf961263c07