Remove V0 workflow. Use CUDA 11.3 for V1 workflow. (#596)
Summary:
CUDA 10.2 is not supported soon and we are using CUDA 11.3 for both V1 and V2 scores.
We are also upgrading to Python 3.8.

This PR also removes V0 workflows, as V0 will not be maintained (e.g. won't backport the BC-breaking torchvision/torchtext fixes to V0).

Upgrading to CUDA 11.3 may cause score change in V1 score, and we will monitor if any regression/optimization is overshadowed by the bump of the CUDA version.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/596

Reviewed By: bitfort

Differential Revision: D32773979

Pulled By: xuzhao9

fbshipit-source-id: 893798422b6c26d3efb30c5e29f99d459baf3d80