Remove the maml models as they have poor quality and should be replaced. (#736)
Summary:
This PR removes two models, maml and maml_omiglot. We are working on another round of enhancing models with tensorrt and other features, since these two models are in low quality and bad gpu utilization, and it is not worth to work on these models. So I am removing them for now.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/736

Reviewed By: erichan1

Differential Revision: D34018844

Pulled By: xuzhao9

fbshipit-source-id: 70da2086bf9782a351f9f8e79238d4d2029229ad