Checkout the main branch before finding the correct commit on date. (#804)
Summary:
Recent bisection job fails because of a bug in the bisection script.
Before building the dependencies, it tries to search the correct commit on that date. However, we need to first checkout the main branch before using that commit. Otherwise, it will only find the commit based on the previous run.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/804

Test Plan: Bisection CI

Reviewed By: erichan1

Differential Revision: D34925803

Pulled By: xuzhao9

fbshipit-source-id: e19dbe43d4384743b9391faf0db20b0749996130