Delete dead Py2 code from pytorch/benchmark (#866)
Summary:
Pull Request resolved: https://github.com/pytorch/benchmark/pull/866

This code was found to be Python 2 only, meaning it cannot run and is polluting
our codebase. This is an automatic codemod to clean it up.

If you have no use for this code, please "Accept & Ship".

If you still use this code via an unsupported Python 2 runtime, please migrate to Python 3.
The Python 2 runtimes will not be around for long, and running Python 2 code is considered
a security vulnerability.

If this code is unused we ask that you do not keep it around for any historical purpose.
It is not cost-free to keep this code, as it breaks codemods, linters, and confuses n00bs.

Reviewed By: xuzhao9

Differential Revision: D35618187

fbshipit-source-id: 61b34a23b2d596b835b073a5742b1d99b21d27fe