Do not report CPU dispatching time in run.py (#880)
Summary:
We should not report CPU dispatch time in run.py because it's inaccurate for many models, such as timm.
This is because in third-party model libraries, it is common to have cuda sync in the middle of model execution, in which case we can't really calculate CPU dispatch time by substracting `t1-t0`.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/880

Reviewed By: aaronenyeshi

Differential Revision: D35952468

Pulled By: xuzhao9

fbshipit-source-id: 5b036b5392e14f675c3782e5308ad1676d0b7b35