Archive fambench_dlrm due to slow time-to-signal (#1107)
Summary:
fambench_dlrm is too slow and large for torchbench, the model is >13G and takes a long time to initialize. Yet it still represents some of our internal workloads, so we still keep it in a separate archive directory.
Plan: the models in `archived_models` can be run with explicit calls with `run.py`, but they won't be visible in sweep runs.

Fixes https://github.com/pytorch/benchmark/issues/1106 .

Pull Request resolved: https://github.com/pytorch/benchmark/pull/1107

Reviewed By: jansel

Differential Revision: D38686136

Pulled By: xuzhao9

fbshipit-source-id: 7fde7c1fa6ef4b563275f0d7dbf56ca4727f4600