Remove incorrect MAML implementation
Fixes https://github.com/pytorch/benchmark/issues/328

Here's some context:
- We discovered that the implementation doesn't actually use maml
(https://github.com/dragen1860/MAML-Pytorch/issues/59)
- We filed https://github.com/pytorch/benchmark/issues/328
- We added maml_omniglot https://github.com/pytorch/benchmark/pull/349
as the correct version of of the maml model.
- We didn't delete the maml model (because I was worried that it was
doing a "different" type of maml that I hadn't seen before that is still
valid).

The last step to resolve this issue is to delete the incorrect MAML
example, unless we have reasons to keep it around.