Improve bisection workflow (#1210)
Summary:
This PR improves the bisection workflow by moving the GitHub issue part to the bisection workflow.
After the bisection finishes, if the result is non-empty, create a GitHub issue tracker. Otherwise, treat it as noise and do not create an issue.

TODO: we still need to post GitHub issues when bisection workflow fails, but we can leave it to a follow-up PR. See https://github.com/pytorch/benchmark/issues/1212

Pull Request resolved: https://github.com/pytorch/benchmark/pull/1210

Reviewed By: erichan1

Differential Revision: D39827220

Pulled By: xuzhao9

fbshipit-source-id: 77ea2bb8d742035cd9c90aee2b7c6f56f265f9ea