[DO NOT MERGE] Required changes after cuBLAS workspace allocation change
The cuBLAS workspace PR https://github.com/pytorch/pytorch/pull/85447 moves allocations that were previously managed by cuBLAS to the `CUDACachingAllocator`. While the lifetime of this memory has not changed, it is now visible to `torch.cuda.memory_allocated()` which can throw false-positives in testing results. A workaround to get an accurate memory measurement is to clear all allocated cuBLAS workspaces via `torch._C._cuda_clearCublasWorkspaces()` which was also added in this PR.

Do not merge this for now as the cuBLAS workspace PR has been reverted.

CC @xuzhao9 @seemethere