Allow equal_nan in correctness check
Summary:
For some models it's not straightforward to generate random inputs
that avoid NaNs in the output.  Since NaNs always compare false, our
correctness checks on the output of an optimized version of the model will
appear to fail, despite producing the same result as the original model.

To avoid such spurious failure we add an `EQUAL_NAN` flag to the model, and
propagate it to the output correctness check.

Reviewed By: jspark1105

Differential Revision: D40372371

fbshipit-source-id: d3481ec5af53431d8075b04bc10e4e6a1612334c