Update on "ddp_experiments: improve error handling"
- Add a timeout on the model subprocess itself; this way, the subprocess
will get killed and we won't timeout on the outer barrier call
- When waiting for the model subprocess, wait in a loop where we
alternate between checking the queue and checking if the process is
finished. This prevents us from waiting for the entire timeout when the
worker process fails, and also prevents the worker process from hanging
in the case where the queue fills up and blocks.

Differential Revision: [D40825608](https://our.internmc.facebook.com/intern/diff/D40825608)

[ghstack-poisoned]