Clarify instllation step
The value of <benchmark> used in "git clone" and "cd" are not the same.
For git clone, https://github.com/pytorch/benchmark is required to use
and then it's required to cd to "benchmark" directory.
Also, providing repo link for git clone is a standard practice and
can be easy to understand for someone totally new to PyTorch and
benchmark.