Clarify instllation step (#1426)
Summary:
The value of `<benchmark>` used in `git clone` and `cd` steps are not the same. For `git clone`, https://github.com/pytorch/benchmark is required to use and then it's required to `cd` to `benchmark` directory.
Also, providing a repo link for git clone is a standard practice, and it can be easy to understand for someone new to PyTorch and benchmark.
This PR provides a small improvement towards clarifying it.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/1426

Reviewed By: aaronenyeshi

Differential Revision: D43735698

Pulled By: xuzhao9

fbshipit-source-id: a35850f99d6f1b8231e7f8f63bc63398280c6fc0