Fix the Inductor speedup benchmark
Summary:
Fixes a bug in the internal inductor_speedup benchmark.

Starting from March 17, 2023, the inductor_speedup benchmark on AIBench failed because of timeout.

The reason is because the introduction of "static_extension_finder" module, which runs in the prelude of the python process. We need to specify the extra system path in the child process environment variable `PYTHONPATH`.

Reviewed By: jianyuh

Differential Revision: D44352332

fbshipit-source-id: 2a4b00deb506aac88ea6571d8db76173d4da4535