Incorporate errors.txt into the issue alert (#1564)
Summary:
This way, both runtime and perf regressions can be detected. Note that this only checks that the errors log file exists without reading from it.

This assumes that the errors.txt file will be generated by the benchmarks run step in the nightly and allows for users to tweak the path if desired.

DISCLAIMER: if the benchmarks fail due to an unrecoverable error and do not run til completion (such as OOMing), this will still NOT catch the issue, since it requires complete logs to be generated and exist.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/1564

Reviewed By: xuzhao9

Differential Revision: D45192313

Pulled By: janeyx99

fbshipit-source-id: 4dff5e0fc4b78da059db8c67dd45ed2369dccf29