replace np.int with int [2]
Summary:
`numpy.int` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `int` and should be safe to replace with `int`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.int\b' | xargs perl -pi -e 's,\bnp\.int\b,int,g'
```

Differential Revision: D46607067

fbshipit-source-id: 1b5390746ec05e5c2f581975c6c43f91f34e4480