Get PT-D benchmark to support being run internally
Summary:
Pytorch Distributed benchmarks have a hard dependency on `submitit` which is not supported internally (we don't use slurm).

It outputs this error: `ModuleNotFoundError: No module named 'submitit'`

These changes remove that hard dependency and add a scheduler option and allow it to be run internally and sets up the code for future work.
The scheduler defaults to slurm so there is no change, but this will eventually be updated to default to local.

Reviewed By: davidberard98

Differential Revision: D47682027

fbshipit-source-id: 9197e2ecb58d507236707b0be7a6bb2dea496269