the accuracy test always returns none for the output
Summary:
As titled, we should fix it to have the actual output return and compared.

Before the fix, test will always pass because it is comparing None with None

After the fix, it can compare the actual outputs.

we detected output mismatch in cmf_10x, cmf_10x_batch_fusion and ads_dhen_5x and alexnet now (see v2), which we will disable and need investigation ore replacing with up-to-date models.

Reviewed By: xuzhao9, jackiexu1992

Differential Revision: D49077949

fbshipit-source-id: 2952bf007e522bf8cacc1415b92349a82d1fab9b