Log non-pt2_compliant ops encountered by Dynamo
Summary:
X-link: https://github.com/pytorch/pytorch/pull/112581

See internal diff for more changes. Whenever we encounter a non-compliant op,
we add it to a set on the OutputGraph. When a compilation event happens, we log
the contents of this set.

I'm planning on flipping the `only_allow_pt2_compliant_ops` config from False
to True after the logging determines that existing models do not use
non-compliant ops.

Reviewed By: yanboliang

Differential Revision: D50884828

fbshipit-source-id: 9866322c19b6b5cf702ac8b2afb2f450f0b9b438