Add manual bisector
Summary:
Sometime we observe performance signals from the AIBench dashboard, it does not block the Diff landing or trunk health, but we would like to know if we can find the root cause commit of this change locally on a devgpu.

This diff adds a manual bisector which runs on a local devgpu server to run the bisection automatically. It requires `fbclone` and `buck` tools, so it can't be deployed to remote tupperware like AIBench for now.

The usage is:
```
buck2 run mode/opt //pytorch/benchmark:bisection -- <userbenchmark-name> --config <bisection-config>
```

Where the `<userbenchmark-name>` is the name of the torchbench userbenchmark you would like to bisect. And `<bisection-config>` specifies the details of the bisection, such as starting commit hash, ending commit hash, and the userbenchmark metrics users would like to bisect.

It has the following features:
1. During bisection, it will only bisect on the *specified* regressed metrics, instead of running the entire benchmark suite. For example, if only 1 model regressed out of 10, it will only run that single model.
2. It supports bisecting both numeric metrics and string metrics. If numeric, it calculates the delta value between 2 commits. If string, it does byte-wise comparison to determine a regression.

Reviewed By: aaronenyeshi

Differential Revision: D50228997

fbshipit-source-id: 72ea77bf63690f0b16c7f6948f4c31f0fd41e90a