Support tensors as Dict keys Re-PR of #111196
Summary:
This prepares the PR where we implement sets in terms of dicts.
To do so, rather than storing internally a dictionary that maps literals
to VariableTrackers, it stores (pretty much) a dictionary from VTs to VTs.
To do so, keys are wrapped in an opaque internal class _Hashable.
The Hashable class is opaque on purpose so that it fails hard if
if it inadvertently leaks back into user code.

We also found and fixed a number of latent bugs and inconsistencies
in the way dynamo checked what can be a dict key. More generally, we
make much clearer what are the things that need to be modified to add
a new supported key type to Dicts.

Fixes #107595
Fixes #111603

Re-PR of https://github.com/pytorch/pytorch/pull/111196 - sadly due to reverts, we could not reuse lezcano's original PR.

X-link: https://github.com/pytorch/pytorch/pull/116785

Reviewed By: jansel, mlazos

Differential Revision: D52546937

Pulled By: voznesenskym

fbshipit-source-id: f82b977ab110b54aeaa802858107fda24f9ee7f9