Improve bisection routine by creating GitHub issue only when the bisection is successful (#2050)
Summary:
This is to resolve the noisy issues of the V3 nightly workflow.
When a regression is detected, we will first kick off a bisection workflow, and only create the GitHub issue when the bisection workflow fails or finds the root cause commit. If the bisection succeeds, but it does not find any result, we will not create an issue.

Pull Request resolved: https://github.com/pytorch/benchmark/pull/2050

Test Plan: https://github.com/pytorch/benchmark/actions/runs/7506472200

Reviewed By: aaronenyeshi

Differential Revision: D52782455

Pulled By: xuzhao9

fbshipit-source-id: 61a8e68a26643ec0a16ab83e1f123102faf8de3c