disable torch autograd for eval test mode
Tests for torch dynamo with inductor backend are
not invoking the FX pass in Torch compilation because
the auto grad is not disabled. This commit is disabling
the auto grad globally for eval tests where there is
no grad calculations are required.

For example, for this test
python run_benchmark.py cpu --model hf_Bert --test eval --torchdynamo inductor