Fix header sorting, off-by-one error in range
Summary:
The headers were sometimes not aligned with the values, because the
headers weren't sorted and the metrics were.  I tidied up the logic (IMO) for
the headers to make it a bit clearer what was happening.

I also noticed an off by one error in how the range of x-vals was computed, so
I rolled that fix in too.

Reviewed By: xuzhao9

Differential Revision: D56574195

fbshipit-source-id: efab091f4b8455e0f899bb6d2a02dc84fb5d16e3