Add has_guarded_code to CompilationMetrics (#125279)
Summary:
While studying some tlparse, I noticed that CompilationMetrics was reporting that there was no error for frames that have no nodes. I'm pretty sure we don't actually install a frame in this situation. has_guarded_code will tell us if that's the case, because it says if the GuardedCode object is None or not.

Actually, while working on this, I was wondering if we can ever trigger the "skip this frame entirely, do not trace it ever again" codepath, as best as I could tell, it's impossible for this to happen by the time we get to compilation metrics block.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

X-link: https://github.com/pytorch/pytorch/pull/125279
Approved by: https://github.com/yanboliang

Reviewed By: kit1980

Differential Revision: D56851167

Pulled By: ezyang

fbshipit-source-id: cdd4380407d414196d14c3239740b397d4ce49db