Scale GEMM inputs for better accuracy results
Summary:
This provides more numerically stable inputs for GEMMs. The +1
eliminates very small values that could result in denormals, and the
scale (which should be set to K in an M*N*K GEMM) means that the
elements of the output tensor are close to integer values. bwasti came
up with this.

Reviewed By: bertmaher

Differential Revision: D57084742

fbshipit-source-id: 13ea8b72f0a502db0e37bd5a92fc12fd3e7688d6