Fix a bug in collecting NCU replay file
Summary:
davidberard98 discovered a bug where we are using a hack by renaming the `ncu_rep` metric to `ncu_trace` to skip adding a new metric name. This is incorrect when users are collecting ncu replay files for multiple inputs.

In this Diff, we fix this by adding a new field for the NCU replay file in the built-in metrics.

Reviewed By: davidberard98

Differential Revision: D58208171

fbshipit-source-id: 88568bb6e3536d3830474dbdb25453d215aab269