Add support for reducing across the middle dimension for 3D matrices using the sum Triton kernel (#2297)
Summary:
Pull Request resolved: https://github.com/pytorch/benchmark/pull/2297

Support reducing 3-dimensional matrices across the middle dimension (`dim == 1`) such that the result is of dimensions `(M, K)`. This kernel assumes that `BLOCK_SIZE_M == 1`, as Triton is currently unable to perform reductions on a middle dimension, and that that the entire reduction dimension of the tensor fits in a thread block (`BLOCK_SIZE_N >= N`).

Reviewed By: davidberard98

Differential Revision: D58307854

fbshipit-source-id: 77a4225cad80c1f3ac6d6355f4c5e0e221e62ff5