Add package contraints to torchbench (#2318)
Summary:
This PR introduces package constraints because we want to guarantee that certain packages are the same version before and after the installation. This is important for the CI in pytorch/pytorch because overriding critical package versions will break the CI due to ABI incompatibility.

If we find a certain model breaks the constraints in the future, we could ask the model installer to apply the `constraint.txt` file with their requirements.txt file.

PyTorch CI Test: https://github.com/pytorch/pytorch/pull/129114

Pull Request resolved: https://github.com/pytorch/benchmark/pull/2318

Reviewed By: huydhn

Differential Revision: D58801449

Pulled By: xuzhao9

fbshipit-source-id: 13e9c258f04fecd2b14e5a236c4c07ce54873d9b