Add plots variable along 1 input parameter to jagged_sum operator in TritonBench
Summary:
Add plotting functionality to the `jagged_sum` operator in TritonBench, enabling the creation of line plots for any set of benchmarks variable along one of the following input parameters: `B`, `M`, `seqlen`, or `sparsity`. This diff makes it easier to visualize the differences in `latency` among the different benchmarks in the `jagged_sum` operator.

Add a command-line argument to toggle the benchmarks displayed on the plots, making it easier to visualize just the 2 Triton benchmarks or just the 2 PyTorch benchmarks, if necessary. This modification helps more clearly visualize the `latency` differences between the simple fused and variable-length loop Triton kernels as well as the unpadded and padded PyTorch benchmarks.

Note that for plots testing multiple values of `B` or `M`, the x-axis is on a log scale to more accurately depict trends in latency; whether or not the x-axis is on a log scale is noted in the plot name in the Test Plan.

Reviewed By: davidberard98

Differential Revision: D59034792

fbshipit-source-id: d27e8e59856d24328c25192069d0afbc655a62e7