logging hardware specific data in userbench
Summary:
currently we are saving the device type in this table ie cuda vs rocm vs cpu etc. We want to further update to allow for hardware specific data ie a100 vs h100. Here we are specifying this for cuda based devices only.

If using cuda we will dynamically find the gpu model via nvidia-smi and log that model.

in this case for H100 we will see "Nvidia H100" and for a100 we will see "NVidia A100"

this will allow us to run the same benchmark on both types of gpu and find the difference in the results quickly via scuba at any given time

Differential Revision: D61229059